+++ app/src/main/java/com/example/vrach/ui/util/AppBar.kt
@@ -1,0 +1,44 @@
+package com.example.vrach.ui.util
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import com.example.vrach.VrachScreen
+
+@Composable
+fun VrachAppBar(
+    currentScreen: VrachScreen,
+    canNavigateBack: Boolean,
+    navigateUp: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+
+    TopAppBar(
+        title = {
+            if (currentScreen.name != "Login") {
+                Text(currentScreen.title)
+            }
+        },
+        colors = TopAppBarDefaults.mediumTopAppBarColors(
+            containerColor = Color.White
+        ),
+        modifier = modifier,
+        navigationIcon = {
+            if (canNavigateBack) {
+                IconButton(onClick = navigateUp) {
+                    Icon(
+                        imageVector = Icons.Filled.ArrowBack,
+                        contentDescription = "a"
+                    )
+                }
+            }
+        }
+    )
+}
